requirements.txt
numpy
pyzmq
mypy
matplotlib
scipy
tensorflow
